fixed mulle__genericpointerset_for * added mulle_set_intersect (and mulle__pointerset_release_all_generic) for set intersection * added mulle_map_is_empty and mulle_map_remove_all * added mulle_pointerarray_remove and mulle_array_remove_unique (finally) reduce pedantically some warnings

+++ src/assoc/mulle-assoc.c
@@ -138,8 +138,8 @@ void   _mulle_assoc_remap_intptr_key_range( struct mulle_assoc *assoc,
 
    for( ; p < sentinel; p++)
    {
-      value  = (intptr_t) p->key;
-      if( value >= end)
+      value = (intptr_t) p->key;
+      if( (uintptr_t) value >= end)
          break;
 
       value += offset;
@@ -229,8 +229,8 @@ void   _mulle_assoc_move_intptr_key_range( struct mulle_assoc *assoc,
 
    for( ; p < sentinel; p++)
    {
-      value  = (intptr_t) p->key;
-      if( value > end)
+      value = (intptr_t) p->key;
+      if( (uintptr_t) value > end)
          break;
 
       if( mulle_range_contains_location( move_away_range, value))
